PanelError

+++ EndPoint.Site/Utilities/ClaimUtility.cs
@@ -8,12 +8,12 @@ namespace EndPoint.Site.Utilities
 {
     public static class ClaimUtility
     {
-        public static long? GetUserId(ClaimsPrincipal User)
+        public static int? GetUserId(ClaimsPrincipal User)
         {
             try
             {
                 var claimsIdentity = User.Identity as ClaimsIdentity;
-                long userId = long.Parse(claimsIdentity.FindFirst(ClaimTypes.NameIdentifier).Value);
+                int userId = int.Parse(claimsIdentity.FindFirst(ClaimTypes.NameIdentifier).Value);
                 return userId;
             }
             catch (Exception)
